requirements.txt
amqp==2.1.4
appdirs==1.4.0
billiard==3.5.0.2
celery==4.0.2
Django==1.10.5
django-filter==1.0.1
django-rest-swagger==0.3.8
djangorestframework==3.5.4
kombu==4.0.2
Markdown==2.6.8
packaging==16.8
psycopg2==2.6.2
pyparsing==2.1.10
pytz==2016.10
PyYAML==3.12
six==1.10.0
vine==1.1.3
